Move space info into the only feature using it
