+++ app/lib/features/space/widgets/space_info.dart
